requirements.txt
inquirer
coloroma
gTTS
Pillow
pytesseract
pdf2image